Renames legal name form container to be more explicit
